chore(wmg): move CellInfoSideBar to wmg2 folder
