+++ frontend/src/views/WheresMyGene/components/Main/index.tsx
@@ -27,7 +27,7 @@ import {
 } from "../../common/store/actions";
 import { ChartProps, GeneExpressionSummary } from "../../common/types";
 import { SideBarPositioner, SideBarWrapper, Top, Wrapper } from "../../style";
-import CellInfoBar from "../CellInfoSideBar";
+import CellInfoBar from "../../../WheresMyGeneV2/components/CellInfoSideBar";
 import GeneInfoBar from "../../../../components/GeneInfoSideBar";
 import Filters from "../Filters";
 import GetStarted from "../GetStarted";
@@ -39,7 +39,7 @@ import ScreenTint from "../ScreenTint";
 import { StyledBannerContainer, StyledSidebarDrawer } from "./style";
 import RightSideBar from "../../../../components/common/RightSideBar";
 import BottomBanner from "src/components/BottomBanner";
-import { CELL_INFO_SIDEBAR_WIDTH_PX } from "../CellInfoSideBar/style";
+import { CELL_INFO_SIDEBAR_WIDTH_PX } from "../../../WheresMyGeneV2/components/CellInfoSideBar/style";
 import { GENE_EXPRESSION_BANNER_SURVEY_LINK } from "src/common/constants/airtableLinks";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport";
 import GeneSearchBar from "src/views/WheresMyGeneV2/components/GeneSearchBar";
