+++ frontend/src/views/WheresMyGeneV2/components/CellInfoSideBar/constants.ts
