+++ frontend/src/views/WheresMyGeneV2/components/CellInfoSideBar/style.ts
