+++ frontend/src/views/WheresMyGeneV2/components/CellInfoSideBar/types.ts
@@ -1,6 +1,6 @@
 import { RightSidebarProperties } from "../../../../components/common/RightSideBar";
 import { OntologyTerm } from "src/common/queries/wheresMyGene";
-import { State } from "../../common/store";
+import { State } from "../../../WheresMyGene/common/store";
 
 export interface CellInfoBarProps extends RightSidebarProperties {
   cellInfoCellType: Exclude<State["cellInfoCellType"], null>;
